groovy if

+++ Jenkinsfile
@@ -3,33 +3,70 @@ pipeline {
     stages {
         stage('Build') {
             steps {
-                sh '''
-                docker build -t stratcastor/duo-deploy-flask:latest -t stratcastor/duo-deploy-flask:v$BUILD_NUMBER .
-                '''
+                script {
+                    if (env.GIT_BRANCH == 'dev') {
+                        sh '''
+                        docker build -t stratcastor/duo-deploy-flask:latest -t stratcastor/duo-deploy-flask:v$BUILD_NUMBER .
+                        '''
+                    } else {
+                        sh '''
+                        echo "Build Not required"
+                        '''
+                    }
+                }
             }
         }
         stage('Push') {
             steps {
-                sh '''
-                docker push stratcastor/duo-deploy-flask:latest
-                docker push stratcastor/duo-deploy-flask:v$BUILD_NUMBER
-                '''
+                script {
+                    if (env.GIT_BRANCH == 'dev') {
+                        sh '''
+                        docker push stratcastor/duo-deploy-flask:latest
+                        docker push stratcastor/duo-deploy-flask:v$BUILD_NUMBER
+                        '''
+                    } else {
+                        sh '''
+                        echo "Push Not required"
+                        '''
+                    }
+                }
             }
         }
         stage('Cleanup') {
             steps {
-                sh '''
-                docker rmi stratcastor/duo-deploy-flask:latest
-                docker rmi stratcastor/duo-deploy-flask:v$BUILD_NUMBER
-                '''
+                script {
+                    if (env.GIT_BRANCH == 'dev') {
+                        sh '''
+                        docker rmi stratcastor/duo-deploy-flask:latest
+                        docker rmi stratcastor/duo-deploy-flask:v$BUILD_NUMBER
+                        '''
+                    } else {
+                        sh '''
+                        echo "Cleanup Not required"
+                        '''
+                    }
+                }
             }
         }
         stage('Deploy') {
             steps {
-                sh '''
-                kubectl apply -f ./k8s-deployments -n production
-                kubectl rollout restart deployment flask-deployment -n production
-                '''
+                script {
+                    if (env.GIT_BRANCH == 'dev') {
+                        sh '''
+                        kubectl apply -f ./k8s-deployments -n development
+                        kubectl rollout restart deployment flask-deployment -n development
+                        '''
+                    } else if (env.GIT_BRANCH == 'main') {
+                        sh '''
+                        kubectl apply -f ./k8s-deployments -n production
+                        kubectl rollout restart deployment flask-deployment -n production
+                        '''
+                    } else {
+                        sh '''
+                        echo "Branch not recognised"
+                        '''
+                    }
+                }
             }
         }
     }
